requirements.txt
streamlit==1.39.0
langchain-core==0.3.13
langchain-community==0.3.3
langchain-groq==0.2.0
chromadb==0.5.16
requests==2.32.3
python-dotenv==1.0.1
pandas
langchain_nomic
nomic[local]